Fixed bug in `parsing terminal inclusion`.
'>' stops the parsing loop and is not pushed into acc(umulator)
we need to add it manually. In the older loop tokenizer branch this was
done just before yielding, because the tag subclass recognition was
faked and only needed to look at first characters.

Now that we used .. heavier checks (e.g '<!--', '-->') that assumes a
full tag '<...>' with both begin and end brackets. So '>' addition has
to be done before decision.